+++ out/getHtml.js
@@ -145,7 +145,7 @@ function createEditorHtml(context, initialContent) {
 									<span>Has header</span>
 								</label>
 								<span class="tooltip is-tooltip-right is-tooltip-multiline"
-									data-tooltip="The first row is used as header. Note that changing this option will also change the write header option.">
+								data-tooltip="The first row is used as header. Note that changing this option will also change the write header option. It will also clear the undo/redo stack!">
 									<i class="fas fa-question-circle"></i>
 								</span>
 							</div>
@@ -417,7 +417,7 @@ function createEditorHtml(context, initialContent) {
 							</span>
 							<span>Trim</span>
 							<span class="tooltip mar-left-half is-tooltip-multiline is-tooltip-left"
-								data-tooltip="Trims ever cell (including header row) in the table (removes leading and trailing spaces, tabs, ...). This will clear undo/redo!">
+							data-tooltip="Trims ever cell (including header row) in the table (removes leading and trailing spaces, tabs, ...). This will clear undo/redo stack!">
 								<i class="fas fa-question-circle"></i>
 							</span>
 						</button>